refactor of components into component directory
