+++ src/components/EntryModal.jsx
